+++ src/app/[username]/components/profile/ProfileSection.tsx
@@ -4,11 +4,11 @@ import React from 'react';
 
 import { useParams } from 'next/navigation';
 
-import { FollowButton } from '@/components/buttons/followButton';
 import Icon from '@/components/icon';
 import { Button } from '@/components/ui/button';
 import { Separator } from '@/components/ui/separator';
 import { toast } from '@/components/ui/use-toast';
+import { FollowButton } from '@/components/user/buttons/followButton';
 import { useSession } from 'next-auth/react';
 
 import { ConnectionsDialog } from './ConnectionsDialog';
