+++ src/app/[username]/components/profile/TopicsCard.tsx
@@ -3,7 +3,7 @@ import { useParams } from 'next/navigation';
 import LinksRedirectArrow from '@/components/links/LinksRedirectArrow';
 import { Loader } from '@/components/loader';
 import { Badge } from '@/components/ui/badge';
-import useUser from '@/hooks/useUser';
+import useUser from '@/hooks/user/useUser';
 import { useSession } from 'next-auth/react';
 
 export const TopicsCard = () => {
