+++ src/app/edit/[blogId]/page.tsx
@@ -13,8 +13,8 @@ import PublishModal from '@/components/modals/publish/PublishModal';
 import { Button } from '@/components/ui/button';
 import { toast } from '@/components/ui/use-toast';
 import { getEditorConfig } from '@/config/editor/editorjs.config';
-import useGetDraftBlogDetail from '@/hooks/useGetDraftBlogDetail';
-import useGetPublishedBlogDetailByBlogId from '@/hooks/useGetPublishedBlogDetailByBlogId';
+import useGetDraftBlogDetail from '@/hooks/blog/useGetDraftBlogDetail';
+import useGetPublishedBlogDetailByBlogId from '@/hooks/blog/useGetPublishedBlogDetailByBlogId';
 import axiosInstance from '@/services/api/axiosInstance';
 import { OutputData } from '@editorjs/editorjs';
 import { useSession } from 'next-auth/react';
