+++ src/app/explore-topics/components/CategoryButton.tsx
@@ -5,7 +5,7 @@ import Icon from '@/components/icon';
 import { Loader } from '@/components/loader';
 import { Button } from '@/components/ui/button';
 import { useToast } from '@/components/ui/use-toast';
-import useUser from '@/hooks/useUser';
+import useUser from '@/hooks/user/useUser';
 import axiosInstance from '@/services/api/axiosInstance';
 import { useSession } from 'next-auth/react';
 import { mutate } from 'swr';
