+++ src/app/explore-topics/components/TopicsList.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 
 import { Button } from '@/components/ui/button';
-import useUser from '@/hooks/useUser';
+import useUser from '@/hooks/user/useUser';
 import { useSession } from 'next-auth/react';
 
 import { TopicButton } from './TopicButton';
